Homework 2 finished
Implement scatter-linear hierarchy generator.

* The implementation takes advantage of "ordinary" C++03 TypeList. Every such list is supposed to end with NullType.
* For purposes of beauty, alias construction (from C++11) is used in GenScatterF. It can be safely removed from the code, so that the whole program satisfies C++03 standard

For now, the applications of classes created using the "library" are unclear. A quite useless example of GenScatterF is provided in main.cpp (this is for testing purposes; the underlying code is compiled when "no optimizations" directive is given to compiler. Extra testing cases is a topic to be dicsussed.