Homework 3 (#3)
* Homework 3 version 1

This is a "8-mark" implementation.

The implementation has no special features (not described in the task).

However, it has to overcome the incompleteness of description in
how to determine how many bytes should be read before passing them to decompressor class.

This is done by an introduction of a special static member in Decompressor class. This is a weird, but simple and straightforward solution.

In the possible commits to follow, this should be replaced by a Decompressor that takes pointer to file.
If this is done, Decompressor can determine the size it reads from file directly,
taking into consideration possible "divisors" of fields in file.

It is unclear how to implement the "10-mark" implementation right now;
the main reason is that such implementation should take pointers to functions
as template arguments, which does not seem quite correct.

* fixup! Homework 3 version 1

Provide full and correct implementation of "8-mark" solution

The Decompressor class now takes pointer to file and can read as many bytes as it finds necessary from there.

Also, a separate "template-branch" is provided for "5-mark" solution (read uncompressed data from file and pass no compressor classes)

* Homework 3 version 2

This is a "10-mark" implementation

Provide full implementation for a list of decompressors.
There are a few features (and limitations) not mentioned in the task. They are as follows:
* The user should use functors as function wrappers. They are defined in DecompressionFunctors.h.
	* Note that there are two functors in that file; their description contains information on when to use which
* Decompression functions (wrapped inside functors) MUST meet one criterion:
	* Use the same buffer for input and output
	This implies, however, that AT ANY STAGE, "uncompressed" output MUST NOT be greater than "compressed" input.
	We use destination buffer as a temporary storage.
	Functions that do not meet this requirement may cause SEGFAULT.

Support for "older" APIs (with just one compressor) is the same as before.